+++ cognite/experimental/_api/functions.py
@@ -332,11 +332,25 @@ def _validate_function_handle(function_handle):
 
 
 class FunctionCallsAPI(APIClient):
-    def list(self, function_id: Optional[int] = None, function_external_id: Optional[str] = None) -> FunctionCallList:
-        """`List all calls associated with a specific function. <https://docs.cognite.com/api/playground/#operation/get-api-playground-projects-project-functions-function_name-calls>`_
+    _LIST_CLASS = FunctionCallList
+
+    def list(
+        self,
+        function_id: Optional[int] = None,
+        function_external_id: Optional[str] = None,
+        status: Optional[str] = None,
+        schedule_id: Optional[int] = None,
+        start_time: Optional[Dict[str, int]] = None,
+        end_time: Optional[Dict[str, int]] = None,
+    ) -> FunctionCallList:
+        """`List all calls associated with a specific function id. Either function_id or function_external_id must be specified. <https://docs.cognite.com/api/playground/#operation/get-api-playground-projects-project-functions-function_name-calls>`_
 
         Args:
             function_id (int, optional): ID of the function on which the calls were made.
+            status (str, optional): Status of the call. Possible values ["Running", "Failed", "Completed", "Timeout"].
+            schedule_id (int, optional): Schedule id from which the call belongs (if any).
+            start_time (Union[Dict[str, int], TimestampRange]): Start time of the call. Possible keys are `min` and `max`, with values given as time stamps in ms.
+            end_time (Union[Dict[str, int], TimestampRange]): End time of the call. Possible keys are `min` and `max`, with values given as time stamps in ms.
             function_external_id (str, optional): External ID of the function on which the calls were made.
 
         Returns:
@@ -361,8 +375,10 @@ class FunctionCallsAPI(APIClient):
         utils._auxiliary.assert_exactly_one_of_id_or_external_id(function_id, function_external_id)
         if function_external_id:
             function_id = self._cognite_client.functions.retrieve(external_id=function_external_id).id
-        url = f"/functions/{function_id}/calls"
-        res = self._get(url)
+        url = f"/functions/{function_id}/calls/list"
+        filter = {"status": status, "scheduleId": schedule_id, "startTime": start_time, "endTime": end_time}
+        post_body = {"filter": filter}
+        res = self._post(url, json=post_body)
         return FunctionCallList._load(res.json()["items"], cognite_client=self._cognite_client)
 
     def retrieve(