+++ CHANGELOG.md
@@ -12,8 +12,10 @@ Changes are grouped as follows
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities.
 
-
 ## Unreleased
+### Changed
+- Function response is no longer a property of the `FunctionCall` class. Instead, the response can be retrieved by the methods `FunctionCallsAPI.get_response()` or `FunctionCall.get_response()`.
+
 ### Added
 - Filtering of function calls given call attributes and added  an attribute for schedule id to the FunctionCall data class.